+++ src/steps/index.ts
@@ -1,6 +1,7 @@
 import { organizationSteps } from './organization';
 import { userSteps } from './user';
+import { cloudWAFSteps } from './cloudwaf';
 
-const integrationSteps = [...organizationSteps, ...userSteps];
+const integrationSteps = [...organizationSteps, ...userSteps, ...cloudWAFSteps];
 
 export { integrationSteps };